Refactoring of feature dto, separating responsibilities and improving code readability.
